improves how we sort stuff
